# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Experiments API",
    "description": "Experiment APIs to create experiment model, managed by the Cosmos team. Reach out in the #experiments channel on Slack for support.",
    "version": "v4"
  },
  "host": "api01-np.agro.services",
  "schemes": [
    "https"
  ],
  "basePath": "/experiments-api",
  "paths": {
...
    "/experiments": {
      "get": {
        "summary": "get all experiment entities",
        "description": "This endpoint returns information about the experiments.\n",
        "operationId": "getExperiments",
        "tags": [
          "Experiments"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tags.category",
            "in": "query",
            "description": "Tag categories comma separated list",
            "required": false,
            "type": "string"
          },
          {
            "name": "tags.value",
            "in": "query",
            "description": "Tag values comma separated list",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination argument to limit the amount of returned results",
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination argument to offset the returned results",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of experiments",
            "schema": {
              "$ref": "#/definitions/PaginatedExperimentResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "summary": "post one or more experiment entities",
        "description": "Creates new experiment(s) in the Experiments System. The definition to use for the POST body is based upon the source query parameter as described below: <ul><li>template: <pre>{\n  \"id\": 0,\n  \"numberOfCopies\": 0,\n  \"name\": \"new name\"\n}</pre><br/></li><li>experiment: <pre>{\n  \"ids\": [\n    0\n  ],\n  \"numberOfCopies\": 0\n}</pre><br/></li><li>undefined/not provided: <pre>[\n  {\n    \"name\": \"string\",\n    \"description\": \"string\",\n    \"experimentDesign\": \"string\",\n    \"status\": \"string\",\n    \"tags\": [\n      {\n        \"name\": \"string\",\n        \"value\": \"string\",\n        \"experimentId\": 0\n      }\n    ],\n    \"owners\": [\n      \"string\"\n    ],\n    \"ownerGroups\": [\n      \"string\"\n    ],\n    \"reviewers\": [\n      \"string\"\n    ],\n  \"reviewerUsers\": [\n      \"string\"\n    ],\n \"request\": {\n      \"id\": \"string\",\n      \"type\": \"string\"\n    }\n  }\n]</pre></li></ul>",
        "operationId": "postExperiments",
        "tags": [
          "Experiments"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "source",
            "type": "string",
            "in": "query"
          },
          {
            "name": "username",
            "description": "An optional header to allow APIs to indicate which user they are modifying data on behalf of. This is only required for API-to-API calls.",
            "in": "header",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "Experiment Created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CreateResponse"
              }
            }
          },
          "400": {
            "description": "Invalid/Problematic Input",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/experiments/{id}": {
      "get": {
        "summary": "get one experiment entity",
        "description": "The Experiments endpoint returns information about the experiment entities in Experiments System.\n",
        "operationId": "getExperimentById",
        "tags": [
          "Experiments"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment entity",
            "schema": {
              "$ref": "#/definitions/ExperimentResponse"
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No Experiment Found For ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "put one experiment entity",
        "description": "update experiment entity in Experiments System.\n",
        "operationId": "putExperimentById",
        "tags": [
          "Experiments"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "experiment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ExperimentPayload"
            }
          },
          {
            "name": "username",
            "description": "An optional header to allow APIs to indicate which user they are modifying data on behalf of. This is only required for API-to-API calls.",
            "in": "header",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment entity",
            "schema": {
              "$ref": "#/definitions/ExperimentResponse"
            }
          },
          "400": {
            "description": "Invalid/Problematic Input",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No Experiment Found For ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "summary": "deletes a single experiment",
        "description": "This endpoint allows a user to delete a experiment",
        "operationId": "deleteExperiment",
        "tags": [
          "Experiments"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "username",
            "description": "An optional header to allow APIs to indicate which user they are modifying data on behalf of. This is only required for API-to-API calls.",
            "in": "header",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment Deleted",
            "schema": {
              "$ref": "#/definitions/GenericDeleteSuccess"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/experiments/{id}/blocks": {
      "patch": {
        "summary": "rename blocks for experiment",
        "description": "Renames the specified blocks that belong to the experiment.",
        "operationId": "renameExperimentsBlocks",
        "tags": [
          "Blocks"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "blockRenames",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BlockRenameRequest"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success with no content"
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No Template Found For ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/experiments/{id}/capacity-request-sync": {
      "post": {
        "summary": "Synchronizes capacity request data with an experiment's data",
        "description": "Synchronizes specific fields from a request from the Capacity Request system with the data stored for an Experiment",
        "operationId": "capacityRequestSync",
        "tags": [
          "Capacity"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "Capacity Request Sync Values",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "locations": {
                  "type": "integer"
                },
                "number of rows": {
                  "type": "integer"
                },
                "plot row length uom": {
                  "type": "string"
                },
                "reps": {
                  "type": "integer"
                },
                "row length": {
                  "type": "integer"
                },
                "row spacing": {
                  "type": "integer"
                },
                "row spacing uom": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "username",
            "description": "An optional header to allow APIs to indicate which user they are modifying data on behalf of. This is only required for API-to-API calls.",
            "in": "header",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "204": {
            "description": "Sync Successful"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/experiments/{id}/response-variables": {
      "get": {
        "summary": "get all response variables for given experiment id.",
        "description": "This endpoint returns information about the response variable entities associated with one Experiment.\n",
        "operationId": "getResponseVariablesByExperimentId",
        "tags": [
          "Response Variables"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Response Variables",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResponseVariableResponse"
              }
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "No Experiment Found For ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/experiments/{id}/permissions": {
      "get": {
        "summary": "get user permissions for the user for experiment id",
        "description": "This endpoint returns information about the user permissions om experiment",
        "operationId": "getPermissionsForUserByExperimentId",
        "tags": [
          "Permissions"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PermissionsResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BlockRenameRequest": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id of the block to rename."
          },
          "name": {
            "type": "string",
            "description": "The new name of the block."
          }
        }
      }
    },
...
    "CreateResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "description": "http status code, 201 for success and 4.x.x for errors."
        },
        "message": {
          "type": "string",
          "description": "Resource created or error description in case of error."
        },
        "id": {
          "type": "integer",
          "description": "id of newly created resource or black in case of error."
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "code": {
          "type": "string"
        },
        "errorCode": {
          "type": "string"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
...
    "PaginatedExperimentResponse": {
      "type": "object",
      "properties": {
        "pageResults": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExperimentResponse"
          },
          "description": "array of paginated results subject to offset/limit"
        },
        "totalResultsLength": {
          "type": "integer",
          "description": "The amount of results prior to applying pagination"
        },
        "pageResultsLength": {
          "type": "integer",
          "description": "The amount of results after applying pagination"
        }
      }
    },
...
    "GenericDeleteSuccess":{
      "type": "object",
      "properties":{
        "id": {
          "type": "integer",
          "description": "Experiment/Template id"
        }
      }
    },
...
  }
}